Implement add user login data to DB (user sign up).
